+++ backend/canisters/group/impl/src/lib.rs
@@ -7,7 +7,7 @@ use activity_notification_state::ActivityNotificationState;
 use candid::Principal;
 use canister_state_macros::canister_state;
 use canister_timer_jobs::TimerJobs;
-use chat_events::Reader;
+use chat_events::{GroupChatThreadKeyPrefix, KeyPrefix, Reader};
 use event_store_producer::{EventStoreClient, EventStoreClientBuilder, EventStoreClientInfo};
 use event_store_producer_cdk_runtime::CdkRuntime;
 use fire_and_forget_handler::FireAndForgetHandler;
@@ -366,6 +366,12 @@ impl RuntimeState {
                 now,
             );
         }
+        for (message_index, _) in result.threads {
+            self.data
+                .stable_memory_keys_to_garbage_collect
+                .push(KeyPrefix::GroupChatThread(GroupChatThreadKeyPrefix::new(message_index)));
+        }
+        jobs::garbage_collect_stable_memory::start_job_if_required(self);
     }
 
     pub fn metrics(&self) -> Metrics {
@@ -420,6 +426,7 @@ impl RuntimeState {
                 .unwrap_or_default(),
             event_store_client_info: self.data.event_store_client.info(),
             timer_jobs: self.data.timer_jobs.len() as u32,
+            stable_memory_event_migration_complete: self.data.stable_memory_event_migration_complete,
             canister_ids: CanisterIds {
                 user_index: self.data.user_index_canister_id,
                 group_index: self.data.group_index_canister_id,
@@ -471,6 +478,10 @@ struct Data {
     expiring_member_actions: ExpiringMemberActions,
     user_cache: UserCache,
     user_event_sync_queue: GroupedTimerJobQueue<UserEventBatch>,
+    #[serde(default)]
+    stable_memory_event_migration_complete: bool,
+    #[serde(default)]
+    stable_memory_keys_to_garbage_collect: Vec<KeyPrefix>,
 }
 
 fn init_instruction_counts_log() -> InstructionCountsLog {
@@ -566,6 +577,8 @@ impl Data {
             expiring_member_actions: ExpiringMemberActions::default(),
             user_cache: UserCache::default(),
             user_event_sync_queue: GroupedTimerJobQueue::new(5, true),
+            stable_memory_event_migration_complete: true,
+            stable_memory_keys_to_garbage_collect: Vec::new(),
         }
     }
 
@@ -725,6 +738,7 @@ pub struct Metrics {
     pub serialized_chat_state_bytes: u64,
     pub event_store_client_info: EventStoreClientInfo,
     pub timer_jobs: u32,
+    pub stable_memory_event_migration_complete: bool,
     pub canister_ids: CanisterIds,
 }