+++ backend/canisters/group/impl/src/lifecycle/init.rs
@@ -1,4 +1,5 @@
 use crate::lifecycle::{init_env, init_state};
+use crate::memory::get_chat_events_memory;
 use crate::Data;
 use canister_tracing_macros::trace;
 use group_canister::init::Args;
@@ -11,6 +12,7 @@ use utils::env::Environment;
 #[trace]
 fn init(args: Args) {
     canister_logger::init(args.test_mode);
+    chat_events::ChatEvents::init_stable_storage(get_chat_events_memory());
 
     let mut env = init_env([0; 32]);