+++ backend/canisters/user/impl/src/jobs/mod.rs
@@ -1,0 +1,7 @@
+use crate::RuntimeState;
+
+pub mod garbage_collect_stable_memory;
+
+pub(crate) fn start(state: &RuntimeState) {
+    garbage_collect_stable_memory::start_job_if_required(state);
+}