+++ backend/integration_tests/src/communities/mod.rs
@@ -2,6 +2,7 @@ mod access_gate_expiry_tests;
 mod cancel_invites_tests;
 mod convert_group_into_community_tests;
 mod create_channel_tests;
+mod delete_channel_tests;
 mod disappearing_message_tests;
 mod import_group_tests;
 mod join_channel_tests;