+++ backend/libraries/chat_events/src/chat_events.rs
@@ -2,6 +2,8 @@ use crate::chat_events_list::Reader;
 use crate::expiring_events::ExpiringEvents;
 use crate::last_updated_timestamps::LastUpdatedTimestamps;
 use crate::search_index::SearchIndex;
+use crate::stable_storage::key::KeyPrefix;
+use crate::stable_storage::Memory;
 use crate::*;
 use candid::Principal;
 use event_store_producer::{EventBuilder, EventStoreClient, Runtime};
@@ -10,22 +12,24 @@ use rand::rngs::StdRng;
 use rand::Rng;
 use search::{Document, Query};
 use serde::{Deserialize, Serialize};
+use serde_bytes::ByteBuf;
 use sha2::{Digest, Sha256};
 use std::cmp::{max, Reverse};
 use std::collections::hash_map::Entry::{Occupied, Vacant};
-use std::collections::{HashMap, HashSet};
+use std::collections::{BTreeMap, HashMap, HashSet};
 use std::mem;
 use std::ops::DerefMut;
+use tracing::info;
 use types::{
     AcceptP2PSwapResult, CallParticipant, CancelP2PSwapResult, CanisterId, Chat, ChatType, CompleteP2PSwapResult,
-    CompletedCryptoTransaction, Cryptocurrency, DirectChatCreated, EventIndex, EventWrapper, EventWrapperInternal,
-    EventsTimeToLiveUpdated, GroupCanisterThreadDetails, GroupCreated, GroupFrozen, GroupUnfrozen, Hash, HydratedMention,
-    Mention, Message, MessageContentInitial, MessageEditedEventPayload, MessageEventPayload, MessageId, MessageIndex,
-    MessageMatch, MessageReport, MessageTippedEventPayload, Milliseconds, MultiUserChat, P2PSwapAccepted, P2PSwapCompleted,
-    P2PSwapCompletedEventPayload, P2PSwapContent, P2PSwapStatus, PendingCryptoTransaction, PollVotes, ProposalUpdate,
-    PushEventResult, Reaction, ReactionAddedEventPayload, RegisterVoteResult, ReserveP2PSwapResult, ReserveP2PSwapSuccess,
-    TimestampMillis, TimestampNanos, Timestamped, Tips, UserId, VideoCall, VideoCallEndedEventPayload, VideoCallParticipants,
-    VideoCallPresence, VoteOperation,
+    CompletedCryptoTransaction, Cryptocurrency, DirectChatCreated, EventContext, EventIndex, EventWrapper,
+    EventWrapperInternal, EventsTimeToLiveUpdated, GroupCanisterThreadDetails, GroupCreated, GroupFrozen, GroupUnfrozen, Hash,
+    HydratedMention, Mention, Message, MessageContentInitial, MessageEditedEventPayload, MessageEventPayload, MessageId,
+    MessageIndex, MessageMatch, MessageReport, MessageTippedEventPayload, Milliseconds, MultiUserChat, P2PSwapAccepted,
+    P2PSwapCompleted, P2PSwapCompletedEventPayload, P2PSwapContent, P2PSwapStatus, PendingCryptoTransaction, PollVotes,
+    ProposalUpdate, PushEventResult, Reaction, ReactionAddedEventPayload, RegisterVoteResult, ReserveP2PSwapResult,
+    ReserveP2PSwapSuccess, TimestampMillis, TimestampNanos, Timestamped, Tips, UserId, VideoCall, VideoCallEndedEventPayload,
+    VideoCallParticipants, VideoCallPresence, VoteOperation,
 };
 
 pub const OPENCHAT_BOT_USER_ID: UserId = UserId::new(Principal::from_slice(&[228, 104, 142, 9, 133, 211, 135, 217, 129, 1]));
@@ -35,7 +39,7 @@ const MEMO_PRIZE_REFUND: [u8; 8] = [0x4f, 0x43, 0x5f, 0x50, 0x52, 0x5a, 0x52, 0x
 pub struct ChatEvents {
     chat: Chat,
     main: ChatEventsList,
-    threads: HashMap<MessageIndex, ChatEventsList>,
+    threads: BTreeMap<MessageIndex, ChatEventsList>,
     metrics: ChatMetricsInternal,
     per_user_metrics: HashMap<UserId, ChatMetricsInternal>,
     frozen: bool,
@@ -45,19 +49,96 @@ pub struct ChatEvents {
     video_call_in_progress: Timestamped<Option<VideoCall>>,
     anonymized_id: String,
     search_index: SearchIndex,
+    #[serde(default = "default_next_event_to_migrate_to_stable_memory")]
+    next_event_to_migrate_to_stable_memory: Option<EventContext>,
+}
+
+fn default_next_event_to_migrate_to_stable_memory() -> Option<EventContext> {
+    Some(EventContext::default())
 }
 
 impl ChatEvents {
+    pub fn init_stable_storage(memory: Memory) {
+        stable_storage::init(memory)
+    }
+
+    pub fn import_events(chat: Chat, events: Vec<(EventContext, ByteBuf)>) {
+        stable_storage::write_events_as_bytes(chat, events);
+    }
+
+    pub fn garbage_collect_stable_memory(prefix: KeyPrefix) -> Result<u32, u32> {
+        stable_storage::garbage_collect(prefix)
+    }
+
+    pub fn set_stable_memory_key_prefixes(&mut self) {
+        self.main.set_stable_memory_prefix(self.chat, None);
+        for (message_index, events) in self.threads.iter_mut() {
+            events.set_stable_memory_prefix(self.chat, Some(*message_index));
+        }
+    }
+
+    pub fn migrate_next_batch_of_events_to_stable_storage(&mut self) -> (usize, bool) {
+        if self.next_event_to_migrate_to_stable_memory.is_none() {
+            return (0, true);
+        };
+
+        let mut total_count = 0;
+        while ic_cdk::api::instruction_counter() < 1_000_000_000 {
+            let EventContext {
+                thread_root_message_index: next_thread_root_message_index,
+                event_index: next_event_index,
+            } = self.next_event_to_migrate_to_stable_memory.clone().unwrap();
+
+            let (thread_root_message_index, events_list) = if let Some(message_index) = next_thread_root_message_index {
+                if let Some((index, next)) = self.threads.range_mut(message_index..).next() {
+                    (Some(*index), next)
+                } else {
+                    self.next_event_to_migrate_to_stable_memory = None;
+                    self.main.set_read_events_from_stable_memory(true);
+                    for events_list in self.threads.values_mut() {
+                        events_list.set_read_events_from_stable_memory(true);
+                    }
+                    info!(chat = ?self.chat, total_count, "Finished migrating events to stable memory");
+                    return (total_count, true);
+                }
+            } else {
+                (None, &mut self.main)
+            };
+
+            let (count, next_event_index) = events_list.migrate_events_to_stable_memory(next_event_index, 100);
+            if let Some(event_index) = next_event_index {
+                self.next_event_to_migrate_to_stable_memory = Some(EventContext {
+                    thread_root_message_index,
+                    event_index,
+                });
+            } else {
+                self.next_event_to_migrate_to_stable_memory = Some(EventContext {
+                    thread_root_message_index: Some(thread_root_message_index.map_or(MessageIndex::default(), |m| m.incr())),
+                    event_index: EventIndex::default(),
+                });
+            }
+            total_count += count;
+        }
+        info!(
+            chat = ?self.chat,
+            count = total_count,
+            next = ?self.next_event_to_migrate_to_stable_memory,
+            "Migrated batch of events to stable memory"
+        );
+        (total_count, false)
+    }
+
     pub fn new_direct_chat(
         them: UserId,
         events_ttl: Option<Milliseconds>,
         anonymized_id: u128,
         now: TimestampMillis,
     ) -> ChatEvents {
+        let chat = Chat::Direct(them.into());
         let mut events = ChatEvents {
-            chat: Chat::Direct(them.into()),
-            main: ChatEventsList::default(),
-            threads: HashMap::new(),
+            chat,
+            main: ChatEventsList::new(chat, None),
+            threads: BTreeMap::new(),
             metrics: ChatMetricsInternal::default(),
             per_user_metrics: HashMap::new(),
             frozen: false,
@@ -67,6 +148,7 @@ impl ChatEvents {
             video_call_in_progress: Timestamped::default(),
             anonymized_id: hex::encode(anonymized_id.to_be_bytes()),
             search_index: SearchIndex::default(),
+            next_event_to_migrate_to_stable_memory: None,
         };
 
         events.push_event(None, ChatEventInternal::DirectChatCreated(DirectChatCreated {}), 0, now);
@@ -83,10 +165,11 @@ impl ChatEvents {
         anonymized_id: u128,
         now: TimestampMillis,
     ) -> ChatEvents {
+        let chat = chat.into();
         let mut events = ChatEvents {
-            chat: chat.into(),
-            main: ChatEventsList::default(),
-            threads: HashMap::new(),
+            chat,
+            main: ChatEventsList::new(chat, None),
+            threads: BTreeMap::new(),
             metrics: ChatMetricsInternal::default(),
             per_user_metrics: HashMap::new(),
             frozen: false,
@@ -96,6 +179,7 @@ impl ChatEvents {
             video_call_in_progress: Timestamped::default(),
             anonymized_id: hex::encode(anonymized_id.to_be_bytes()),
             search_index: SearchIndex::default(),
+            next_event_to_migrate_to_stable_memory: None,
         };
 
         events.push_event(
@@ -114,6 +198,14 @@ impl ChatEvents {
 
     pub fn set_chat(&mut self, chat: Chat) {
         self.chat = chat;
+        self.main.set_stable_memory_prefix(chat, None);
+        for (message_index, events_list) in self.threads.iter_mut() {
+            events_list.set_stable_memory_prefix(chat, Some(*message_index));
+        }
+    }
+
+    pub fn read_events_as_bytes_from_stable_memory(&self, after: Option<EventContext>) -> Vec<(EventContext, ByteBuf)> {
+        stable_storage::read_events_as_bytes(self.chat, after, 1_000_000)
     }
 
     pub fn iter_recently_updated_events(
@@ -122,13 +214,19 @@ impl ChatEvents {
         self.last_updated_timestamps.iter()
     }
 
+    pub fn thread_keys(&self) -> impl Iterator<Item = MessageIndex> + '_ {
+        self.threads.keys().copied()
+    }
+
     pub fn push_message<R: Runtime + Send + 'static>(
         &mut self,
         args: PushMessageArgs,
         mut event_store_client: Option<&mut EventStoreClient<R>>,
     ) -> EventWrapper<Message> {
         let events_list = if let Some(root_message_index) = args.thread_root_message_index {
-            self.threads.entry(root_message_index).or_default()
+            self.threads
+                .entry(root_message_index)
+                .or_insert_with(|| ChatEventsList::new(self.chat, Some(root_message_index)))
         } else {
             &mut self.main
         };
@@ -1643,7 +1741,10 @@ impl ChatEvents {
         correlation_id: u64,
         now: TimestampMillis,
     ) -> EventIndex {
-        let events = self.threads.entry(thread_root_message_index).or_default();
+        let events = self
+            .threads
+            .entry(thread_root_message_index)
+            .or_insert_with(|| ChatEventsList::new(self.chat, Some(thread_root_message_index)));
         events.push_event(event, correlation_id, None, now)
     }