+++ backend/libraries/types/src/event_index.rs
@@ -3,6 +3,9 @@ use serde::{Deserialize, Serialize};
 use std::fmt::{Display, Formatter};
 use ts_export::ts_export;
 
+pub const MIN_EVENT_INDEX: EventIndex = EventIndex(0);
+pub const MAX_EVENT_INDEX: EventIndex = EventIndex(u32::MAX);
+
 #[ts_export]
 #[derive(CandidType, Serialize, Deserialize, Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct EventIndex(u32);