+++ backend/canisters/local_group_index/CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add an error log with http endpoint ([#6608](https://github.com/open-chat-labs/open-chat/pull/6608))
 
+### Changed
+
+- Maintain set of which canisters have not yet migrated all events to stable memory ([#6603](https://github.com/open-chat-labs/open-chat/pull/6603))
+
 ## [[2.0.1383](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1383-local_group_index)] - 2024-10-11
 
 ### Added