+++ backend/canisters/user/CHANGELOG.md
@@ -15,10 +15,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Make searching messages more efficient ([#6612](https://github.com/open-chat-labs/open-chat/pull/6612))
 - Add `message_id` and `event_index` to `MessageActivityEvent` ([#6623](https://github.com/open-chat-labs/open-chat/pull/6623))
+- Copy chat events into stable memory ([#6603](https://github.com/open-chat-labs/open-chat/pull/6603))
 
 ### Fixed
 
 - Fix removing link previews ([#6633](https://github.com/open-chat-labs/open-chat/pull/6633))
+- Determine whether c2c call should be retried based on response error ([#6640](https://github.com/open-chat-labs/open-chat/pull/6640))
 
 ## [[2.0.1395](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1395-user)] - 2024-10-16